requirements.txt
Flask>=2.0
Pillow>=9.0
numpy>=1.24
scipy>=1.10
opencv-python>=4.8
gunicorn>=21.2.0
